requirements.txt
av==9.2.0
firebase_admin==5.2.0
Flask==1.1.2
geocoder==1.38.1
geopy==2.2.0
numpy==1.19.5
opencv_python==4.5.5.62
Pillow==9.1.1
requests==2.22.0
streamlit==1.7.0
streamlit_option_menu==0.3.2
streamlit_webrtc==0.37.0
torch==1.10.2
twilio==7.9.2
Werkzeug==2.0.3
